business_facing/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpx==0.27.0
idna==3.10
openrouteservice==2.3.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
